\docType{package}
\name{OLScurve-package}
\alias{OLScurve-package}
\title{OLS growth curve trajectories}
\description{
  OLS growth curve trajectories
}
\details{
  Provides tools for organizing, calculating, and plotting
  ordinary least squares growth curve trajectories. These
  type of models are typically used to diagnose the
  effectiveness of a specified time functional form at the
  individual level rather than at the group level (where
  FIML estimation is required).
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com} and
  Carrie Smith \email{carrie.elizabeth@gmail.com} and
  Matthew Sigal \email{matthewsigal@gmail.com}
}
\keyword{package}

