\docType{data}
\name{gender}
\alias{gender}
\title{Description of gender}
\description{
  A vector identifying whether the repeated measure row was
  Male or Female for the \code{nonlin.example} dataset.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
\keyword{datasets}

