\docType{data}
\name{nonlin.example}
\alias{nonlin.example}
\title{Description of nonlin.example data}
\arguments{
  \item{gender}{a vector identifying whether the repeated
  measure row was Male or Female}
}
\description{
  An artificial dataset simulated by the parameters given
  in the vignette file for \code{OLScurve}.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
\keyword{datasets}

